+++ packages/core/src/components/index.ts
@@ -59,10 +59,6 @@ export * from "./spinner/spinner";
 export * from "./spinner/svgSpinner";
 export * from "./tabs/tab";
 export * from "./tabs/tabs";
-export * from "./tabs/tabList";
-export * from "./tabs/tabPanel";
-export * from "./tabs2/tab2";
-export * from "./tabs2/tabs2";
 export * from "./tag/tag";
 export * from "./tag-input/tagInput";
 export * from "./toast/toast";
