+++ packages/core/src/components/tabs/tabTitle.tsx
@@ -9,9 +9,9 @@ import * as PureRender from "pure-render-decorator";
 import * as React from "react";
 
 import * as Classes from "../../common/classes";
-import { ITab2Props, TabId } from "./tab2";
+import { ITabProps, TabId } from "./tab";
 
-export interface ITabTitleProps extends ITab2Props {
+export interface ITabTitleProps extends ITabProps {
     /** Handler invoked when this tab is clicked. */
     onClick: (id: TabId, event: React.MouseEvent<HTMLElement>) => void;
