+++ packages/core/test/tabs/tabsTests.tsx
@@ -3,19 +3,27 @@
  * Licensed under the terms of the LICENSE file distributed with this project.
  */
 
-// tslint:disable max-classes-per-file
-
 import { assert } from "chai";
 import { mount, ReactWrapper } from "enzyme";
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { spy } from "sinon";
 
-import * as Errors from "../../src/common/errors";
+import * as Classes from "../../src/common/classes";
 import * as Keys from "../../src/common/keys";
-import { Tab, TabList, TabPanel, Tabs } from "../../src/index";
+import { Tab } from "../../src/components/tabs/tab";
+import { ITabsProps, ITabsState, Tabs } from "../../src/components/tabs/tabs";
 
 describe("<Tabs>", () => {
+    const ID = "tabsTests";
+    // default tabs content is generated from these IDs in each test
+    const TAB_IDS = ["first", "second", "third"];
+
+    // selectors using ARIA role
+    const TAB = "[role='tab']";
+    const TAB_LIST = "[role='tablist']";
+    const TAB_PANEL = "[role='tabpanel']";
+
     let testsContainerElement: HTMLElement;
 
     beforeEach(() => {
@@ -25,295 +33,335 @@ describe("<Tabs>", () => {
 
     afterEach(() => testsContainerElement.remove());
 
-    it("renders its template", () => {
-        const wrapper = mount(<Tabs>{getTabsContents()}</Tabs>);
-
-        assert.lengthOf(wrapper.find(Tab), 3);
-        assert.lengthOf(wrapper.find(TabList), 1);
-        assert.lengthOf(wrapper.find(TabPanel), 1);
+    it("gets by without children", () => {
+        assert.doesNotThrow(() => mount(<Tabs id="childless" />));
     });
 
-    it("throws error if TabList is not the first child", () => {
-        assert.throws(
-            () =>
-                mount(
-                    <Tabs>
-                        <span />
-                    </Tabs>,
-                ),
-            Errors.TABS_FIRST_CHILD,
+    it("supports non-existent children", () => {
+        assert.doesNotThrow(() =>
+            mount(
+                <Tabs id={ID}>
+                    {null}
+                    <Tab id="one" />
+                    {undefined}
+                    <Tab id="two" />
+                </Tabs>,
+            ),
         );
     });
 
-    it("throws error if number of Tabs and TabPanels don't match", () => {
-        assert.throws(
-            () =>
-                mount(
-                    <Tabs>
-                        <TabList>
-                            <Tab>foo</Tab>
-                            <Tab>bar</Tab>
-                        </TabList>
-                        <TabPanel />
-                    </Tabs>,
-                ),
-            Errors.TABS_MISMATCH,
+    it("default selectedTabId is first non-null Tab id", () => {
+        const wrapper = mount(
+            <Tabs id={ID}>
+                {null}
+                {<button id="btn" />}
+                {getTabsContents()}
+            </Tabs>,
         );
+        assert.lengthOf(wrapper.find(TAB), 3);
+        assert.strictEqual(wrapper.state("selectedTabId"), TAB_IDS[0]);
+    });
+
+    it("renders one TabTitle for each Tab", () => {
+        const wrapper = mount(<Tabs id={ID}>{getTabsContents()}</Tabs>);
+        assert.lengthOf(wrapper.find(TAB), 3);
+    });
+
+    it("renders all Tab children, but active is not aria-hidden", () => {
+        const activeIndex = 1;
+        const wrapper = mount(<Tabs id={ID}>{getTabsContents()}</Tabs>);
+        wrapper.setState({ selectedTabId: TAB_IDS[activeIndex] });
+        const tabs = wrapper.find(TAB_PANEL);
+        assert.lengthOf(tabs, 3);
+        for (let i = 0; i < TAB_IDS.length; i++) {
+            // hidden unless it is active
+            assert.equal(tabs.at(i).prop("aria-hidden"), i !== activeIndex);
+        }
+    });
+
+    it("renders without `pt-large` when by default", () => {
+        const wrapper = mount(<Tabs id={ID}>{getTabsContents()}</Tabs>);
+        assert.lengthOf(wrapper.find(`.${Classes.TAB_LIST}.${Classes.LARGE}`), 0);
+    });
+
+    it("renders using `pt-large` when large={true}", () => {
+        const wrapper = mount(
+            <Tabs id={ID} large={true}>
+                {getTabsContents()}
+            </Tabs>,
+        );
+        assert.lengthOf(wrapper.find(`.${Classes.TAB_LIST}.${Classes.LARGE}`), 1);
+    });
+
+    it("renderActiveTabPanelOnly only renders active tab panel", () => {
+        const wrapper = mount(
+            <Tabs id={ID} renderActiveTabPanelOnly={true}>
+                {getTabsContents()}
+            </Tabs>,
+        );
+        for (const selectedTabId of TAB_IDS) {
+            wrapper.setState({ selectedTabId });
+            assert.lengthOf(wrapper.find("strong"), 1);
+        }
+    });
+
+    it("sets aria-* attributes with matching IDs", () => {
+        const wrapper = mount(<Tabs id={ID}>{getTabsContents()}</Tabs>);
+        wrapper.find(TAB).forEach(title => {
+            // title "controls" tab element
+            const titleControls = title.prop("aria-controls");
+            const tab = wrapper.find(`#${titleControls}`);
+            // tab element "labelled by" title element
+            assert.isTrue(tab.is(TAB_PANEL), "aria-controls isn't TAB_PANEL");
+            assert.deepEqual(tab.prop("aria-labelledby"), title.prop("id"), "mismatched IDs");
+        });
+    });
+
+    it("clicking selected tab still fires onChange", () => {
+        const tabId = TAB_IDS[0];
+        const changeSpy = spy();
+        const wrapper = mount(
+            <Tabs defaultSelectedTabId={tabId} id={ID} onChange={changeSpy}>
+                {getTabsContents()}
+            </Tabs>,
+            { attachTo: testsContainerElement },
+        );
+        findTabById(wrapper, tabId).simulate("click");
+        assert.isTrue(changeSpy.calledWith(tabId, tabId));
+    });
+
+    it("clicking nested tab should not affect parent", () => {
+        const changeSpy = spy();
+        const wrapper = mount(
+            <Tabs id={ID} onChange={changeSpy}>
+                {getTabsContents()}
+                <Tabs id="nested">
+                    <Tab id="last" title="Click me" />
+                </Tabs>
+            </Tabs>,
+            { attachTo: testsContainerElement },
+        );
+        assert.equal(wrapper.state("selectedTabId"), TAB_IDS[0]);
+        // last Tab is inside nested
+        wrapper
+            .find(TAB)
+            .last()
+            .simulate("click");
+        assert.equal(wrapper.state("selectedTabId"), TAB_IDS[0]);
+        assert.isTrue(changeSpy.notCalled, "onChange invoked");
     });
 
     it("changes tab focus when arrow keys are pressed", () => {
         const wrapper = mount(
-            <Tabs>
-                <TabList>
-                    <Tab>foo</Tab>
-                    <Tab isDisabled={true}>bar</Tab>
-                    <Tab>bax</Tab>
-                    <Tab isDisabled={true}>bar</Tab>
-                </TabList>
-                <TabPanel />
-                <TabPanel />
-                <TabPanel />
-                <TabPanel />
+            <Tabs id={ID}>
+                <Tab id="first" title="First" panel={<Panel title="first" />} />,
+                <Tab disabled={true} id="second" title="Second" panel={<Panel title="second" />} />,
+                <Tab id="third" title="Third" panel={<Panel title="third" />} />,
             </Tabs>,
             { attachTo: testsContainerElement },
         );
 
-        const tabs = testsContainerElement.queryAll(".pt-tab");
-        const tab0 = tabs[0];
-        const tab2 = tabs[2];
-        (tab0 as HTMLElement).focus();
-        wrapper.simulate("keydown", { target: tab0, which: Keys.ARROW_RIGHT });
-        assert.equal(tab2, document.activeElement);
+        const tabList = wrapper.find(TAB_LIST);
+        const tabElements = testsContainerElement.queryAll(TAB);
+        (tabElements[0] as HTMLElement).focus();
 
-        wrapper.simulate("keydown", { target: tab2, which: Keys.ARROW_RIGHT });
-        assert.equal(tab2, document.activeElement);
-
-        wrapper.simulate("keydown", { target: tab2, which: Keys.ARROW_LEFT });
-        assert.equal(tab0, document.activeElement);
-
-        wrapper.simulate("keydown", { target: tab0, which: Keys.ARROW_LEFT });
-        assert.equal(tab0, document.activeElement);
+        tabList.simulate("keydown", { which: Keys.ARROW_RIGHT });
+        assert.equal(document.activeElement, tabElements[2], "move right and skip disabled");
+        tabList.simulate("keydown", { which: Keys.ARROW_RIGHT });
+        assert.equal(document.activeElement, tabElements[0], "wrap around to first tab");
+        tabList.simulate("keydown", { which: Keys.ARROW_LEFT });
+        assert.equal(document.activeElement, tabElements[2], "wrap around to last tab");
+        tabList.simulate("keydown", { which: Keys.ARROW_LEFT });
+        assert.equal(document.activeElement, tabElements[0], "move left and skip disabled");
     });
 
-    it("only renders one TabPanel at a time", () => {
-        const wrapper = mount(<Tabs>{getTabsContents()}</Tabs>);
-        // always renders exactly one TabPanel
-        for (let i = 0; i < 3; i++) {
-            wrapper.setState({ selectedTabIndex: i });
-            assert.lengthOf(wrapper.find(TabPanel), 1);
-        }
+    it("enter and space keys click focused tab", () => {
+        const changeSpy = spy();
+        const wrapper = mount(
+            <Tabs id={ID} onChange={changeSpy}>
+                {getTabsContents()}
+            </Tabs>,
+            { attachTo: testsContainerElement },
+        );
+        const tabList = wrapper.find(TAB_LIST);
+        const tabElements = testsContainerElement.queryAll(TAB);
+
+        // must target different elements each time as onChange is only called when id changes
+        tabList.simulate("keypress", { target: tabElements[1], which: Keys.ENTER });
+        tabList.simulate("keypress", { target: tabElements[2], which: Keys.SPACE });
+
+        assert.equal(changeSpy.callCount, 2);
+        assert.includeDeepMembers(changeSpy.args[0], [TAB_IDS[1], TAB_IDS[0]]);
+        assert.includeDeepMembers(changeSpy.args[1], [TAB_IDS[2], TAB_IDS[1]]);
+    });
+
+    it("animate=false removes moving indicator element", () => {
+        const wrapper = mount(
+            <Tabs id={ID} animate={false}>
+                {getTabsContents()}
+            </Tabs>,
+        );
+        assert.isUndefined(wrapper.state().indicatorWrapperStyle);
+        assert.equal(wrapper.find(".pt-tab-indicator").length, 0);
+    });
+
+    it("removes indicator element when selected tab is removed", () => {
+        const wrapper = mount(<Tabs id={ID}>{getTabsContents()}</Tabs>);
+        // first tab is selected by default. now remove it.
+        const tabIdsWithoutFirstTab = TAB_IDS.slice(1);
+        wrapper.setProps({ children: getTabsContents(tabIdsWithoutFirstTab) });
+        const indicatorStyle = wrapper.state().indicatorWrapperStyle;
+        assert.deepEqual(indicatorStyle, { display: "none" }, "indicator should be hidden");
+    });
+
+    it("leaves indicator element in place when non-selected tab is removed", () => {
+        const wrapper = mount(<Tabs id={ID}>{getTabsContents()}</Tabs>);
+        // first tab is selected by default. now remove the last one.
+        const lastTabIndex = TAB_IDS.length - 1;
+        const tabIdsWithoutLastTab = TAB_IDS.slice(0, lastTabIndex - 1);
+        wrapper.setProps({ children: getTabsContents(tabIdsWithoutLastTab) });
+        assertIndicatorPosition(wrapper, "first");
     });
 
     describe("when state is managed internally", () => {
-        it("sets initialSelectedTabIndex", () => {
-            const TAB_INDEX_TO_SELECT = 2;
-            const wrapper = mount(<Tabs initialSelectedTabIndex={TAB_INDEX_TO_SELECT}>{getTabsContents()}</Tabs>);
-            assert.isTrue(
-                wrapper
-                    .find(Tab)
-                    .at(TAB_INDEX_TO_SELECT)
-                    .prop("isSelected"),
+        const TAB_ID_TO_SELECT = TAB_IDS[1];
+
+        it("defaultSelectedTabId is initially selected", () => {
+            const wrapper = mount(
+                <Tabs id={ID} defaultSelectedTabId={TAB_ID_TO_SELECT}>
+                    {getTabsContents()}
+                </Tabs>,
             );
+            assert.isTrue(findTabById(wrapper, TAB_ID_TO_SELECT).prop("aria-selected"));
         });
 
-        it("does not reset selected tab to initialSelectedTabIndex after a selection is made", () => {
-            const TAB_INDEX_TO_SELECT = 2;
-            const wrapper = mount(<Tabs initialSelectedTabIndex={1}>{getTabsContents()}</Tabs>);
-            wrapper.ref(`tabs-${TAB_INDEX_TO_SELECT}`).simulate("click");
+        it("unknown tab ID hides moving indicator element", () => {
+            const wrapper = mount(
+                <Tabs id={ID} defaultSelectedTabId="unknown">
+                    {getTabsContents()}
+                </Tabs>,
+            );
+            const style = wrapper.state().indicatorWrapperStyle;
+            assert.deepEqual(style, { display: "none" });
+        });
+
+        it("does not reset selected tab to defaultSelectedTabId after a selection is made", () => {
+            const wrapper = mount(
+                <Tabs id={ID} defaultSelectedTabId={TAB_ID_TO_SELECT}>
+                    {getTabsContents()}
+                </Tabs>,
+            );
+            findTabById(wrapper, TAB_ID_TO_SELECT).simulate("click");
             wrapper.update();
-            assert.isTrue(
-                wrapper
-                    .find(Tab)
-                    .at(TAB_INDEX_TO_SELECT)
-                    .prop("isSelected"),
-            );
+            assert.isTrue(findTabById(wrapper, TAB_ID_TO_SELECT).prop("aria-selected"));
         });
 
         it("invokes onChange() callback", () => {
-            const TAB_INDEX_TO_SELECT = 1;
             const onChangeSpy = spy();
-            const wrapper = mount(<Tabs onChange={onChangeSpy}>{getTabsContents()}</Tabs>);
+            const wrapper = mount(
+                <Tabs id={ID} onChange={onChangeSpy}>
+                    {getTabsContents()}
+                </Tabs>,
+            );
 
-            wrapper.ref(`tabs-${TAB_INDEX_TO_SELECT}`).simulate("click");
+            findTabById(wrapper, TAB_ID_TO_SELECT).simulate("click");
             assert.isTrue(onChangeSpy.calledOnce);
-            // previous selection is 0
-            assert.isTrue(onChangeSpy.calledWith(TAB_INDEX_TO_SELECT, 0));
+            // initial selection is first tab
+            assert.isTrue(onChangeSpy.calledWith(TAB_ID_TO_SELECT, TAB_IDS[0]));
+        });
+
+        it("moves indicator as expected", () => {
+            const wrapper = mount(<Tabs id={ID}>{getTabsContents()}</Tabs>);
+            assertIndicatorPosition(wrapper, TAB_IDS[0]);
+
+            wrapper.setProps({ selectedTabId: TAB_ID_TO_SELECT });
+            assertIndicatorPosition(wrapper, TAB_ID_TO_SELECT);
         });
     });
 
-    it("clicking nested tab should not affect parent", () => {
-        const wrapper = mount(
-            <Tabs initialSelectedTabIndex={0}>
-                <TabList>
-                    <Tab>React</Tab>
-                    <Tab>Angular</Tab>
-                    <Tab>Ember</Tab>
-                </TabList>
-                <TabPanel>
-                    <Tabs>{getTabsContents()}</Tabs>
-                </TabPanel>
-                <TabPanel>
-                    <Tabs>{getTabsContents()}</Tabs>
-                </TabPanel>
-                <TabPanel>
-                    <Tabs>{getTabsContents()}</Tabs>
-                </TabPanel>
-            </Tabs>,
-            { attachTo: testsContainerElement },
-        );
-        assert.equal(wrapper.state("selectedTabIndex"), 0);
-        // 3 tabs in parent + 3 in child = 6 tabs. click last and verify unchanged
-        wrapper
-            .find(Tab)
-            .last()
-            .simulate("click");
-        assert.equal(wrapper.state("selectedTabIndex"), 0);
-    });
+    describe("when state is managed externally (selectedTabId prop is provided)", () => {
+        const TAB_ID_TO_SELECT = TAB_IDS[1];
+        const SELECTED_TAB_ID = TAB_IDS[2];
 
-    describe("when state is managed externally (selectedTabIndex prop is provided)", () => {
-        it("prefers selectedTabIndex over initialSelectedTabIndex", () => {
+        it("prefers selectedTabId over defaultSelectedTabId", () => {
             const tabs = mount(
-                <Tabs initialSelectedTabIndex={1} selectedTabIndex={0}>
+                <Tabs id={ID} defaultSelectedTabId={TAB_ID_TO_SELECT} selectedTabId={SELECTED_TAB_ID}>
                     {getTabsContents()}
                 </Tabs>,
             );
-            assert.strictEqual(tabs.state("selectedTabIndex"), 0);
+            assert.strictEqual(tabs.state("selectedTabId"), SELECTED_TAB_ID);
         });
 
-        it("defaults to first tab if invalid index provided", () => {
-            const tabs = mount(<Tabs selectedTabIndex={7}>{getTabsContents()}</Tabs>);
+        it("selects nothing if invalid id provided", () => {
+            const tabs = mount(
+                <Tabs id={ID} selectedTabId="unknown">
+                    {getTabsContents()}
+                </Tabs>,
+            );
 
-            assert.strictEqual(tabs.state("selectedTabIndex"), 0);
+            assert.strictEqual(tabs.state("selectedTabId"), "unknown");
+            assert.isFalse(tabs.find("[aria-selected=true]").exists(), "a tab was selected");
         });
 
-        it("invokes onChange() callback", () => {
-            const TAB_INDEX_TO_SELECT = 1;
+        it("invokes onChange() callback but does not change state", () => {
             const onChangeSpy = spy();
             const tabs = mount(
-                <Tabs selectedTabIndex={0} onChange={onChangeSpy}>
+                <Tabs id={ID} selectedTabId={SELECTED_TAB_ID} onChange={onChangeSpy}>
                     {getTabsContents()}
                 </Tabs>,
             );
 
-            tabs.ref(`tabs-${TAB_INDEX_TO_SELECT}`).simulate("click");
+            findTabById(tabs, TAB_ID_TO_SELECT).simulate("click");
             assert.isTrue(onChangeSpy.calledOnce);
             // old selection is 0
-            assert.isTrue(onChangeSpy.calledWith(TAB_INDEX_TO_SELECT, 0));
+            assert.includeDeepMembers(onChangeSpy.args[0], [TAB_ID_TO_SELECT, SELECTED_TAB_ID]);
+            assert.deepEqual(tabs.state("selectedTabId"), SELECTED_TAB_ID);
         });
 
-        it("doesn't switch tabs automatically", () => {
-            const TAB_INDEX_TO_SELECT = 1;
-            const tabs = mount(<Tabs selectedTabIndex={0}>{getTabsContents()}</Tabs>);
-
-            tabs.ref(`tabs-${TAB_INDEX_TO_SELECT}`).simulate("click");
-            assert.strictEqual(tabs.state("selectedTabIndex"), 0);
+        it("state is synced with selectedTabId prop", () => {
+            const tabs = mount(
+                <Tabs id={ID} selectedTabId={SELECTED_TAB_ID}>
+                    {getTabsContents()}
+                </Tabs>,
+            );
+            assert.deepEqual(tabs.state("selectedTabId"), SELECTED_TAB_ID);
+            tabs.setProps({ selectedTabId: TAB_ID_TO_SELECT });
+            assert.deepEqual(tabs.state("selectedTabId"), TAB_ID_TO_SELECT);
         });
 
-        it("does switch tabs if the user hooks up onChange() to do so", () => {
-            const TAB_INDEX_TO_SELECT = 1;
-            const onChangeSpy = spy();
-            class TestComponent extends React.Component<{}, any> {
-                public state = {
-                    mySelectedTab: 0,
-                };
-
-                public render() {
-                    return (
-                        <Tabs selectedTabIndex={this.state.mySelectedTab} onChange={this.handleChange}>
-                            {getTabsContents()}
-                        </Tabs>
-                    );
-                }
-
-                private handleChange = (selectedTabIndex: number) => {
-                    onChangeSpy();
-                    this.setState({ mySelectedTab: selectedTabIndex });
-                };
-            }
-
-            const wrapper = mount(<TestComponent />);
-            wrapper
-                .find(Tab)
-                .at(TAB_INDEX_TO_SELECT)
-                .simulate("click");
-            assert.strictEqual(wrapper.find(TabPanel).text(), "second panel");
-            // ensure only called once (#502)
-            assert.isTrue(onChangeSpy.calledOnce);
-        });
-
-        it("indicator moves correctly if tabs switch externally via the selectedTabIndex prop", done => {
-            const TAB_INDEX_TO_SELECT = 1;
-            const wrapper = mount(<Tabs selectedTabIndex={0}>{getTabsContents()}</Tabs>, {
-                attachTo: testsContainerElement,
-            });
-            wrapper.setProps({ selectedTabIndex: TAB_INDEX_TO_SELECT });
+        it("indicator moves correctly if tabs switch externally via the selectedTabId prop", done => {
+            const wrapper = mount(
+                <Tabs id={ID} selectedTabId={SELECTED_TAB_ID}>
+                    {getTabsContents()}
+                </Tabs>,
+                { attachTo: testsContainerElement },
+            );
+            wrapper.setProps({ selectedTabId: TAB_ID_TO_SELECT });
             // indicator moves via componentDidUpdate
             setTimeout(() => {
-                assertIndicatorPosition(wrapper, TAB_INDEX_TO_SELECT);
-                done();
-            });
-        });
-
-        it("indicator moves correctly if the user switches tabs and Tab children change simulatenously", done => {
-            const TAB_INDEX_TO_SELECT = 1;
-            class TestComponent extends React.Component<{}, any> {
-                public state = {
-                    mySelectedTab: 0,
-                };
-
-                public render() {
-                    return (
-                        <Tabs selectedTabIndex={this.state.mySelectedTab} onChange={this.handleChange}>
-                            {this.children()}
-                        </Tabs>
-                    );
-                }
-
-                private handleChange = (selectedTabIndex: number) => {
-                    this.setState({ mySelectedTab: selectedTabIndex });
-                };
-
-                private children = () => [
-                    <TabList key={0}>
-                        <Tab>{this.state.mySelectedTab === 1 ? "first (unsaved)" : "first"}</Tab>
-                        <Tab>second</Tab>
-                    </TabList>,
-                    <TabPanel key={1} />,
-                    <TabPanel key={2} />,
-                ];
-            }
-            const wrapper = mount(<TestComponent />, { attachTo: testsContainerElement });
-            wrapper
-                .find(Tab)
-                .at(TAB_INDEX_TO_SELECT)
-                .simulate("click");
-            // indicator moves via componentDidUpdate
-            setTimeout(() => {
-                assertIndicatorPosition(wrapper, TAB_INDEX_TO_SELECT);
+                assertIndicatorPosition(wrapper, TAB_ID_TO_SELECT);
                 done();
             });
         });
     });
 
-    function assertIndicatorPosition(wrapper: ReactWrapper<any, {}>, selectedTabIndex: number) {
-        const style = wrapper.find(TabList).props().indicatorWrapperStyle;
-        assert.isDefined(style, "TabList should have a indicatorWrapperStyle prop set");
+    function findTabById(wrapper: ReactWrapper<ITabsProps, {}>, id: string) {
+        // Need this to get the right overload signature
+        // tslint:disable-next-line:no-object-literal-type-assertion
+        return wrapper.find(TAB).filter({ "data-tab-id": id } as React.HTMLAttributes<{}>);
+    }
+
+    function assertIndicatorPosition(wrapper: ReactWrapper<ITabsProps, ITabsState>, selectedTabId: string) {
+        const style = wrapper.state().indicatorWrapperStyle;
+        assert.isDefined(style, "Tabs should have a indicatorWrapperStyle prop set");
         const node = ReactDOM.findDOMNode(wrapper.instance());
-        const expected = (node.queryAll(".pt-tab")[selectedTabIndex] as HTMLLIElement).offsetLeft;
+        const expected = (node.query(`${TAB}[data-tab-id='${selectedTabId}']`) as HTMLLIElement).offsetLeft;
         assert.isTrue(style.transform.indexOf(`${expected}px`) !== -1, "indicator has not moved correctly");
     }
 
-    function getTabsContents(): Array<React.ReactElement<any>> {
-        // keys are just to avoid React warnings; they're not used in tests
-        return [
-            <TabList key={0}>
-                <Tab>first</Tab>
-                <Tab>second</Tab>
-                <Tab>third</Tab>
-            </TabList>,
-            <TabPanel key={1}>first panel</TabPanel>,
-            <TabPanel key={2}>second panel</TabPanel>,
-            <TabPanel key={3}>third panel</TabPanel>,
-        ];
+    function getTabsContents(tabIds: string[] = TAB_IDS): Array<React.ReactElement<any>> {
+        return tabIds.map(id => <Tab id={id} key={id} panel={<Panel title={id} />} title={id} />);
     }
 });
+
+const Panel: React.SFC<{ title: string }> = ({ title }) => <strong>{title} panel</strong>;
